# src/services/api-server/generated/swagger.yaml
openapi: 3.0.0
components:
...
  schemas:
    TodoStatus:
      type: string
      enum:
        - pending
        - in_progress
        - completed
      description: TODO item status
    TodoItem:
      description: TODO item
      properties:
        id:
          type: string
          description: Unique TODO ID
        content:
          type: string
          description: TODO content/description
        status:
          $ref: "#/components/schemas/TodoStatus"
          description: TODO status
      required:
        - id
        - content
        - status
      type: object
      additionalProperties: false
    TodoListResponse:
      description: TODO list response
      properties:
        taskId:
          type: string
          description: Task ID
        todos:
          items:
            $ref: "#/components/schemas/TodoItem"
          type: array
          description: TODO items
        count:
          type: number
          format: double
          description: Total count
        stats:
          properties:
            pending:
              type: number
              format: double
            inProgress:
              type: number
              format: double
            completed:
              type: number
              format: double
            total:
              type: number
              format: double
          required:
            - pending
            - inProgress
            - completed
            - total
          type: object
          description: Statistics
      required:
        - taskId
        - todos
        - count
        - stats
      type: object
      additionalProperties: false
    ApiErrorResponse:
      description: API Error response
      properties:
        error:
          type: string
          description: Error type/category
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        statusCode:
          type: number
          format: double
          description: HTTP status code
      required:
        - error
      type: object
      additionalProperties: false
      example:
        error: Not Found
        message: Profile 'my-profile' not found
        code: PROFILE_NOT_FOUND
    CreateTodoResponse:
      description: Create TODO response
      properties:
        success:
          type: boolean
          enum:
            - true
          nullable: false
        message:
          type: string
        todo:
          $ref: "#/components/schemas/TodoItem"
          description: Created TODO item
      required:
        - success
        - message
        - todo
      type: object
      additionalProperties: false
    CreateTodoRequest:
      description: Create TODO request
      properties:
        content:
          type: string
          description: TODO content/description
        status:
          $ref: "#/components/schemas/TodoStatus"
          description: "Initial status (default: pending)"
        position:
          type: number
          format: double
          description: "Position to insert at (0-based index, default: append to end)"
      required:
        - content
      type: object
      additionalProperties: false
    UpdateTodoResponse:
      description: Update TODO response
      properties:
        success:
          type: boolean
          enum:
            - true
          nullable: false
        message:
          type: string
        todo:
          $ref: "#/components/schemas/TodoItem"
          description: Updated TODO item
      required:
        - success
        - message
        - todo
      type: object
      additionalProperties: false
    UpdateTodoRequest:
      description: Update TODO request
      properties:
        content:
          type: string
          description: Updated content
        status:
          $ref: "#/components/schemas/TodoStatus"
          description: Updated status
      type: object
      additionalProperties: false
    BulkUpdateTodosResponse:
      description: Bulk update TODOs response
      properties:
        success:
          type: boolean
          enum:
            - true
          nullable: false
        message:
          type: string
        todos:
          items:
            $ref: "#/components/schemas/TodoItem"
          type: array
          description: Updated TODO list
      required:
        - success
        - message
        - todos
      type: object
      additionalProperties: false
    BulkUpdateTodosRequest:
      description: Bulk update TODOs request
      properties:
        todos:
          items:
            $ref: "#/components/schemas/TodoItem"
          type: array
          description: Array of TODO items
      required:
        - todos
      type: object
      additionalProperties: false
    ImportTodosResponse:
      description: Import TODOs response (deprecated - use bulk update instead)
      properties:
        success:
          type: boolean
          enum:
            - true
          nullable: false
        message:
          type: string
        todos:
          items:
            $ref: "#/components/schemas/TodoItem"
          type: array
          description: Imported TODO items
      required:
        - success
        - message
        - todos
      type: object
      additionalProperties: false
      deprecated: true
    ImportTodosRequest:
      description: Import TODOs request (deprecated - use bulk update instead)
      properties:
        todos:
          items:
            $ref: "#/components/schemas/TodoItem"
          type: array
          description: Array of TODO items to import
      required:
        - todos
      type: object
      additionalProperties: false
      deprecated: true
    ExportTodosResponse:
      description: Export TODOs response
      properties:
        taskId:
          type: string
          description: Task ID
        todos:
          items:
            $ref: "#/components/schemas/TodoItem"
          type: array
          description: TODO items
        count:
          type: number
          format: double
          description: Total count
        stats:
          properties:
            pending:
              type: number
              format: double
            inProgress:
              type: number
              format: double
            completed:
              type: number
              format: double
            total:
              type: number
              format: double
          required:
            - pending
            - inProgress
            - completed
            - total
          type: object
          description: Statistics
      required:
        - taskId
        - todos
        - count
        - stats
      type: object
      additionalProperties: false
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
info:
  title: Hoody Code API
  version: 1.0.0
  description: REST API for Hoody Code VSCode Extension - Remote control and
    integration interface
  license:
    name: Apache-2.0
  contact:
    name: Hoody Code
    url: https://hoody.com
paths:
  /todos/{taskId}:
    get:
      operationId: GetTodoList
      responses:
        "200":
          description: TODO list retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoListResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Get TODO list for a task
      summary: Get TODO list
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
    post:
      operationId: CreateTodo
      responses:
        "201":
          description: TODO created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTodoResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Create a new TODO item
      summary: Create TODO
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoRequest"
  /todos/{taskId}/{todoId}:
    patch:
      operationId: UpdateTodo
      responses:
        "200":
          description: TODO updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateTodoResponse"
        "404":
          description: Task or TODO not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Update a TODO item
      summary: Update TODO
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
        - description: TODO ID
          in: path
          name: todoId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoRequest"
    delete:
      operationId: DeleteTodo
      responses:
        "204":
          description: TODO deleted
        "404":
          description: Task or TODO not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Delete a TODO item
      summary: Delete TODO
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
        - description: TODO ID
          in: path
          name: todoId
          required: true
          schema:
            type: string
  /todos/{taskId}/bulk:
    post:
      operationId: BulkUpdateTodos
      responses:
        "200":
          description: TODO list replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUpdateTodosResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Replace entire TODO list
      summary: Replace TODO list
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateTodosRequest"
  /todos/{taskId}/import:
    post:
      operationId: ImportTodos
      responses:
        "200":
          description: TODOs imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportTodosResponse"
        "400":
          description: Invalid TODO data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Import TODOs (deprecated - use bulk update instead)
      summary: Import TODOs
      deprecated: true
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportTodosRequest"
  /todos/{taskId}/export:
    get:
      operationId: ExportTodos
      responses:
        "200":
          description: TODOs exported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportTodosResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      description: Export TODO list as JSON
      summary: Export TODO list
      security:
        - bearer: []
      parameters:
        - description: Task ID
          in: path
          name: taskId
          required: true
          schema:
            type: string
...
servers:
  - url: /api/v1/agent
...